requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
httpx
anyio
distro
sqlalchemy